requirements.txt
Pillow>=9.0.0
click>=8.0.0
tqdm>=4.64.0
numpy>=1.21.0
pytest>=7.0.0
pytest-cov>=4.0.0